backend/requirements.txt
flask==2.3.0
flask-cors==4.0.0
anthropic==0.39.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
